refactor(visualization): refactor the chart recommendation module
