+++ src/visualization/index.ts
@@ -2,5 +2,5 @@
  * Visualization module for chart recommendation and GPT-Vis code generation
  */
 
-export { adviseChartType, chartDefinitions } from './advisor';
+export { adviseChartType } from './advisor';
 export { generateVisualizationHTML } from './generator';
